requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
python-dotenv
psycopg2-binary
requests
